Changing Meta image class

+++ modules/src/lv/edi/EDI_3DAtA/imageio/MetaImage.java
@@ -529,7 +529,7 @@ public class MetaImage {
 				}
 				for(int i=0; i<(int)dimSize.get(1,0); i++){
 					for(int j=0; j<(int)dimSize.get(0,0); j++){
-						image.set(i, j,Math.ceil(255*(((float)shortBuffer.get()-min_value)/(max_value-min_value))));
+						image.set(i, j,Math.ceil(255*(((double)shortBuffer.get()-min_value)/(max_value-min_value))));
 					}
 				}
 				break;
